Type safe

+++ MordenDMALibrary/src/device.cc
@@ -11,7 +11,7 @@ auto Device::Read(uint64_t addr, size_t bytes) const
 }
 
 auto Device::Write(uint64_t addr, const std::vector<uint8_t>& data) const
-    -> std::invoke_result_t<decltype(&DMAIO::Write), const DMAIO, int32_t,
+    -> std::invoke_result_t<decltype(&DMAIO::Write), const DMAIO, uint32_t,
                             uint64_t, std::vector<uint8_t>&> {
-  return std::invoke(&DMAIO::Write, io_, static_cast<int32_t>(-1), addr, data);
+  return std::invoke(&DMAIO::Write, io_, -1, addr, data);
 }
